requirements.txt
django
djangorestframework
djangorestframework-oauth
django-oauth-toolkit
django-rest-framework-social-oauth2
django-filter
python-dotenv
markdown
psycopg2
